Create QA Overview component and some clean up
