+++ src/App.css
@@ -12,13 +12,12 @@
 
 .application-container {
   text-align: center;
-  background-color: rgb(238, 246, 242);
+  background-color: #b1de87;
   min-height: 100%;
 }
 
 footer {
   width: 100%;
-  color: #29cc96;
   margin-top: 1rem;
   font-weight: 500;
   padding: 1.25rem 0;
@@ -104,7 +103,7 @@ footer {
 }
 
 .nav li a {
-  color: #29cc96 !important;
+  color: #596f44 !important;
 }
 
 tr th.text-primary {
