+++ src/components/QATool/QAOverview.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+type Props = {};
+
+export const QAOverview = (props: Props) => {
+  return (
+    <div className="title-block">
+      <h2 className="title">The awesome QA Tool</h2>
+      <h3 className="subtitle">Feel free to add new questions</h3>
+    </div>
+  );
+};
