+++ src/context/methods/requestSort.ts
@@ -1,0 +1,7 @@
+export const requestSort = (sortConfig, setSortConfig, key) => {
+  let direction = 'ascending';
+  if (sortConfig && sortConfig.key === key && sortConfig.direction === 'ascending') {
+    direction = 'descending';
+  }
+  setSortConfig({ key, direction });
+};
