+++ src/qnaSampleData.json
@@ -1,0 +1,1 @@
+[{ "question": "How to add a question", "answer": "Just use the form below" }]
