+++ src/types/QAType.ts
@@ -1,0 +1,4 @@
+export type QAType = {
+  question: string;
+  answer: string;
+};
